Move `CardPresentPaymentsModalViewModelWCSettingsWebViewPresenting` to a separate file.
