+++ WooCommerce/Classes/ViewModels/CardPresentPayments/CardPresentModalConnectionFailedUpdateAddress.swift
@@ -1,20 +1,6 @@
+import Combine
 import UIKit
 import Yosemite
-
-import Combine
-
-struct WCSettingsWebViewModel: Identifiable {
-    var id: String {
-        webViewURL.absoluteString
-    }
-
-    let webViewURL: URL
-    let onCompletion: () -> Void
-}
-
-protocol CardPresentPaymentsModalViewModelWCSettingsWebViewPresenting {
-    var webViewModel: AnyPublisher<WCSettingsWebViewModel?, Never> { get }
-}
 
 /// Modal presented when an error occurs while connecting to a reader due to problems with the address
 ///
